src/backend/requirements.txt
langgraph==0.2.3
langchain==0.2.12
langchain-community==0.2.11
langchain-openai==0.1.21
langchainhub==0.1.20
faiss-cpu==1.8.0.post1
passlib===1.7.4
pydantic<2.0.0
python-dotenv==1.0.1
tiktoken==0.7.0
redis==5.0.8
pydub==0.25.1
pypdf==4.3.1
fpdf==1.7.2
fastapi==0.112.0
uvicorn==0.20.0
python-multipart==0.0.9
selenium==4.23.1
webdriver-manager==4.0.2
bs4==0.0.2
lxml==5.3.0